+++ src/main/java/results/repository/GradeRepository.java
@@ -3,6 +3,6 @@ package results.repository;
 import org.springframework.data.repository.CrudRepository;
 import results.model.Grade;
 
-public interface GradeRepository extends CrudRepository<Grade, Long> {
+public interface GradeRepository extends CrudRepository<Grade, Integer> {
 
 }
