+++ src/main/java/results/repository/SettingRepository.java
@@ -3,6 +3,6 @@ package results.repository;
 import org.springframework.data.repository.CrudRepository;
 import results.model.Setting;
 
-public interface SettingRepository extends CrudRepository<Setting, Long> {
+public interface SettingRepository extends CrudRepository<Setting, Integer> {
 
 }
